Fix: SendVerificationCodeUseCase extracted to view model
