--- app/src/main/java/com/ivangvozdev/buyupix/domain/usecase/SendVerificationCodeUseCase.kt
@@ -1,42 +1,0 @@
-package com.ivangvozdev.buyupix.domain.usecase
-
-import androidx.activity.ComponentActivity
-import com.google.firebase.FirebaseException
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.PhoneAuthCredential
-import com.google.firebase.auth.PhoneAuthOptions
-import com.google.firebase.auth.PhoneAuthProvider
-import java.util.concurrent.TimeUnit
-import javax.inject.Inject
-
-class SendVerificationCodeUseCase @Inject constructor(
-    private val auth: FirebaseAuth
-) {
-    operator fun invoke(
-        phoneNumber: String,
-        activity: ComponentActivity,
-        onVerificationCompleted: (PhoneAuthCredential) -> Unit,
-        onVerificationFailed: (FirebaseException) -> Unit,
-        onCodeSent: (String, PhoneAuthProvider.ForceResendingToken) -> Unit
-    ) {
-        val options = PhoneAuthOptions.newBuilder(auth)
-            .setPhoneNumber(phoneNumber)
-            .setTimeout(60L, TimeUnit.SECONDS)
-            .setActivity(activity)
-            .setCallbacks(object : PhoneAuthProvider.OnVerificationStateChangedCallbacks() {
-                override fun onVerificationCompleted(credential: PhoneAuthCredential) {
-                    onVerificationCompleted(credential)
-                }
-
-                override fun onVerificationFailed(e: FirebaseException) {
-                    onVerificationFailed(e)
-                }
-
-                override fun onCodeSent(verificationId: String, token: PhoneAuthProvider.ForceResendingToken) {
-                    onCodeSent(verificationId, token)
-                }
-            })
-            .build()
-        PhoneAuthProvider.verifyPhoneNumber(options)
-    }
-}
